Firefly can now render to an image!
FIrefly can now use the off screen render target to render details to an image. This is been tested with multiple viewports to see if it works with multiple viewports, and thus can be used in VR. For now, we encode everything to a png and save the file as we dont have a way to render it to the screen (and it is not needed in mobile).