Object constructors do not initialize anything now.
The object constructors do not initialize anything now. This is because on the mobile platform, Clang's UB sanitizer throws a lot of errors when we access member variables and member functions in constructors. For this, we now have a dedicated function called initialize which initializes all the objects when called using the create function.